nargo: more compact nargo PkgId
The additional symbols ('#', '/', '.') in the package id make it super annoying
to build packages from the nix cli since the attrnames need to be quoted, which
breaks cli autocomplete.

This diff also passes the relative workspace crate path thru via `path` instead
of trying to parse it from the old pkg id.